threedcitydb_start_db.py now integrates starting the db.
Prior to this version the db had to be started separatly
and then its IP could be used to check the DB. Now, thanks
to WorkflowTemplate output forwarding, starting the DB is
a singe (workflow) template.
Also the cluster defintion can now be imported directly
(its initialization is done behind the scene).